capstone-project/pyproject.toml
[project]
name = "langraph-agents"
version = "0.1.0"
description = "A2A Ticketing Agents with LangChain and Gemini integration"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.0.0",
    "langchain-google-genai>=0.0.7",
    "langchain-community>=0.0.13",
    "langchain-mcp-adapters>=0.0.5",
    "mcp>=0.2.2",
    "requests>=2.31.0",
    "a2a-sdk[http-server]>=0.3.0",
    "uvicorn>=0.34.2",
    "httpx>=0.24.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
